+++ neuralqx/experimental/nn/projectors/stmh/__init__.py
@@ -12,11 +12,16 @@
 #  See the License for the specific language governing permissions and
 #  limitations under the License.
 
-"""
-Experimental VMC drivers in neuraLQX.
-"""
+from ._wrap import SingleTrunkMultiHeadLogPsi
+from ._wrap import STMHHeadView
+from ._wrap import STMHAllHeadsView
+from ._wrap import wrap_trunk_as_stmh
+from ._wrap import make_stmh_head_models
 
-from .mvmc import MultiStateVMC
-from .stmh_multi_vmc import SingleTrunkMultiHeadVMC
-
-__all__ = ["MultiStateVMC", "SingleTrunkMultiHeadVMC"]
+__all__ = [
+    "SingleTrunkMultiHeadLogPsi",
+    "STMHHeadView",
+    "STMHAllHeadsView",
+    "wrap_trunk_as_stmh",
+    "make_stmh_head_models",
+]