+++ neuralqx/experimental/solver/__init__.py
@@ -17,5 +17,6 @@ Experimental end-to-end solvers in neuraLQX.
 """
 
 from .multi_state_solver import MultiSolver
+from .stmh_solver import STMultiSolver
 
-__all__ = ["MultiSolver"]
+__all__ = ["MultiSolver", "STMultiSolver"]