+++ neuralqx/experimental/vqs/__init__.py
@@ -17,6 +17,11 @@ Experimental variational quantum states in neuraLQX.
 """
 
 from .mc.mc_state import MultiMCState
+from .mc.mc_state import STMultiMCState
 from . import mc
 
-__all__ = ["mc", "MultiMCState"]
+__all__ = [
+    "mc",
+    "MultiMCState",
+    "STMultiMCState",
+]